v2/block/item: Make `syscall::Header::num` `usize`

+++ crates/sallyport/src/v2/src/block/item/syscall.rs
@@ -1,10 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use libc::c_long;
-
 #[derive(Clone, Copy, Debug, PartialEq)]
 #[repr(C, align(8))]
 pub struct Header {
     pub item: super::Header,
-    pub num: c_long,
+    pub num: usize,
 }
